+++ src/rand.rs
@@ -1458,6 +1458,9 @@ pub trait ThreadRandGen {
 static_assert_same_layout!(RandState<'_>, randstate_t);
 static_assert_same_layout!(ThreadRandState<'_>, randstate_t);
 
+static_assert_same_size!(RandState, Option<RandState>);
+static_assert_same_size!(ThreadRandState, Option<ThreadRandState>);
+
 // abort functions do not need a wrapper to abort on panic, they never panic and always abort
 unsafe extern "C" fn abort_seed(_: *mut randstate_t, _: *const mpz_t) {
     process::abort();